+++ src/auth/src/supabase_auth/types.py
@@ -39,7 +39,8 @@ Provider = Literal[
     "slack_oidc",
     "spotify",
     "twitch",
-    "twitter",
+    "twitter",  # Uses OAuth 1.0a
+    "x",  # Uses OAuth 2.0
     "workos",
     "zoom",
 ]
@@ -234,7 +235,10 @@ class User(BaseModel):
     updated_at: Optional[datetime] = None
     identities: Optional[List[UserIdentity]] = None
     is_anonymous: bool = False
+    is_sso_user: bool = False
     factors: Optional[List[Factor]] = None
+    deleted_at: Optional[str] = None
+    banned_until: Optional[str] = None
 
 
 class UserAttributes(TypedDict):